requirements.txt
botbuilder-core
aiohttp
flask
